requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.52.0
httpx==0.25.2
faiss-cpu==1.8.0.post1
python-dotenv==1.0.1
